requirements.txt
SQLAlchemy==2.0.39
fastapi==0.115.11
uvicorn==0.34.0
Jinja2==3.1.6
requests==2.32.4
PySocks==1.7.1
python-multipart==0.0.20
itsdangerous==2.2.0
beautifulsoup4
pillow
openai
deepl
anthropic
langdetect
apscheduler
deepdiff
